refactored react code, adding ajax calls

+++ app/api/forum.py
@@ -7,11 +7,11 @@ import time
 
 forum_blueprint = Blueprint('forum', __name__)
 
-# get thread posts for a given page
-@forum_blueprint.route("/forum/<int:page_num>", methods=['GET'])
+# get ALL thread posts
+@forum_blueprint.route("/forum", methods=['GET'])
 @auth.login_required
     def getForum():
-        posts = db.session.query(Post).order_by(-Post.date.desc()).paginate(page_num, 10, False).items
+        posts = db.session.query(Post).order_by(-Post.date.desc()).all()
         threads = [currPost if currPost.thread_id == -1 for currPost in posts]
         return jsonify([post.serialize() for post in threads])
 
@@ -24,8 +24,8 @@ forum_blueprint = Blueprint('forum', __name__)
 
 
 # add post to db (add to parent thread if needed)
-@forum_blueprint.route("/forum", methods=['POST'])
-@forum_blueprint.route("/forum/<int:thread_id>/", methods=['POST'])
+@forum_blueprint.route("/forum/add", methods=['POST'])
+@forum_blueprint.route("/forum/add/<int:thread_id>", methods=['POST'])
 @auth.login_required
     def addPost(thread_id=-1):
 
@@ -43,13 +43,14 @@ forum_blueprint = Blueprint('forum', __name__)
         post.text = text
         post.name = name
         post.date = int(time.time())
+        post.user_id = user_id
         post.thread_id = thread_id
 
         db.session.add(user)
         db.session.commit()
 
 # edit post
-@forum_blueprint.route("/forum/<int:post_id>/", methods=['POST'])
+@forum_blueprint.route("/forum/edit/<int:post_id>", methods=['POST'])
 @auth.login_required
     def editPost():
         if not request.get_json():
@@ -58,18 +59,22 @@ forum_blueprint = Blueprint('forum', __name__)
         data = request.get_json()
         title = data.get('title')
         text = data.get('text')
+        usr_id = data.get('user_id')
 
-        db.session.query().filter_by(id=post_id).update({"title": title, "text": text})
+        db.session.query().filter_by(user_id=usr_id).filter_by(id=post_id).update({"title": title, "text": text})
         db.session.commit()
 
 # delete post + replies
-@forum_blueprint.route("/forum/<int:post_id>/", methods=['POST'])
+@forum_blueprint.route("/forum/delete/<int:post_id>", methods=['POST'])
 @auth.login_required
     def deletePost():
         if not request.get_json():
             abort(400, 'Malformatted JSON request.')
 
-        db.session.query().filter_by(id=post_id).delete()
-        db.session.query().filter_by(thread_id=post_id).delete()
+        data = request.get_json()
+        usr_id = data.get('user_id')
+
+        db.session.query().filter_by(user_id=usr_id).filter_by(id=post_id).delete()
+        db.session.query().filter_by(user_id=usr_id).filter_by(thread_id=post_id).delete()
         db.session.commit()
